requirements.txt
rich
pyyaml
python-dotenv
colorama
requests
scapy
boofuzz
pyserial
pymodbus
coapthon3
boto3
azure-identity
google-cloud-storage
kubernetes
streamlit
pandas
impacket
ldap3
pyasn1
cryptography
selenium
shodan
dnspython
whois
PyQt6
networkx
matplotlib
pefile
capstone
keystone-engine
frida-tools
androguard
paramiko
pyOpenSSL
plistlib
binwalk
pwntools
python-magic
mitreattack-python
sigma
jinja2
gql
openapi-spec-validator
flask
oauthlib
httpx
